some changes to make things simpler
