+++ README_FIRST.txt
@@ -1,0 +1,22 @@
+
+This is a working version of shadow...
+
+
+Compile shadow3: 
+===============
+
+cd src
+make
+
+
+
+Compile the python bindings:
+============================
+
+cd src
+make lib
+cd ..
+python setup.py build
+
+
+
