--- src/fortran/shadow_version.h
@@ -1,45 +1,0 @@
-print *,"+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++"
-print *,"                           compilation settings                    "
-print *,"+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++"
-print *," "
-print *," +++++date (date):"
-print *,"Thu Jan 14 14:04:22 CET 2016"
-print *," "
-print *," +++++Compiler: "
-print *,"gfortran --version:"
-print *,"GNU Fortran (Ubuntu 4.8.4-2ubuntu1~14.04) 4.8.4"
-print *,"Copyright (C) 2013 Free Software Foundation, Inc."
-print *,""
-print *,"GNU Fortran comes with NO WARRANTY, to the extent permitted by law."
-print *,"You may redistribute copies of GNU Fortran"
-print *,"under the terms of the GNU General Public License."
-print *,"For more information about these matters, see the file named COPYING"
-print *,""
-print *," "
-print *," +++++using Unix platform (uname -a): "
-print *,"Linux chac 3.19.0-30-generic #34~14.04.1-Ubuntu SMP Fri Oct 2 22:09:39 UTC 2015 x86_64 x86_64 x86_64 GNU/Linux"
-print *," "
-print *," +++++building sequence (make -n):"
-print *,"make[1]: Entering directory `/users/srio/Oasys/VIRTUAL_ENVIRONMENT/shadow3/src'"
-print *,"./shadow_version.sh gfortran"
-print *,"gfortran -I ./def/ -cpp -fPIC -ffree-line-length-none   -O2 -fomit-frame-pointer -D_COMPILE4NIX -c fortran/shadow_version.f90 "
-print *,"gfortran -I ./def/ -cpp -fPIC -ffree-line-length-none   -O2 -fomit-frame-pointer -D_COMPILE4NIX -c fortran/shadow_globaldefinitions.f90 "
-print *,"gfortran -I ./def/ -cpp -fPIC -ffree-line-length-none   -O2 -fomit-frame-pointer -D_COMPILE4NIX -c fortran/stringio.f90 "
-print *,"gfortran -I ./def/ -cpp -fPIC -ffree-line-length-none   -O2 -fomit-frame-pointer -D_COMPILE4NIX -c fortran/gfile.f90 "
-print *,"gfortran -I ./def/ -cpp -fPIC -ffree-line-length-none   -O2 -fomit-frame-pointer -D_COMPILE4NIX -c fortran/shadow_beamio.f90 "
-print *,"gfortran -I ./def/ -cpp -fPIC -ffree-line-length-none   -O2 -fomit-frame-pointer -D_COMPILE4NIX -c fortran/shadow_math.f90 "
-print *,"gfortran -I ./def/ -cpp -fPIC -ffree-line-length-none   -O2 -fomit-frame-pointer -D_COMPILE4NIX -c fortran/shadow_variables.f90 "
-print *,"gfortran -I ./def/ -cpp -fPIC -ffree-line-length-none   -O2 -fomit-frame-pointer -D_COMPILE4NIX -c fortran/shadow_roughness.f90 "
-print *,"gfortran -I ./def/ -cpp -fPIC -ffree-line-length-none   -O2 -fomit-frame-pointer -D_COMPILE4NIX -c fortran/shadow_kernel.f90 "
-print *,"gfortran -I ./def/ -cpp -fPIC -ffree-line-length-none   -O2 -fomit-frame-pointer -D_COMPILE4NIX -c fortran/shadow_synchrotron.f90 "
-print *,"gfortran -I ./def/ -cpp -fPIC -ffree-line-length-none   -O2 -fomit-frame-pointer -D_COMPILE4NIX -c fortran/shadow_pre_sync.f90 "
-print *,"gfortran -I ./def/ -cpp -fPIC -ffree-line-length-none   -O2 -fomit-frame-pointer -D_COMPILE4NIX -c fortran/shadow_pre_sync_urgent.f90 "
-print *,"gfortran -I ./def/ -cpp -fPIC -ffree-line-length-none   -O2 -fomit-frame-pointer -D_COMPILE4NIX -c fortran/shadow_preprocessors.f90 "
-print *,"gfortran -I ./def/ -cpp -fPIC -ffree-line-length-none   -O2 -fomit-frame-pointer -D_COMPILE4NIX -c fortran/shadow_postprocessors.f90 "
-print *,"gfortran -I ./def/ -cpp -fPIC -ffree-line-length-none   -O2 -fomit-frame-pointer -D_COMPILE4NIX -c fortran/shadow_bind_f.f90 "
-print *,"gfortran -I ./def/ -cpp -fPIC -ffree-line-length-none   -O2 -fomit-frame-pointer -D_COMPILE4NIX -c fortran/shadow_crl.f90 "
-print *,"gfortran -I ./def/ -cpp -fPIC -ffree-line-length-none   -O2 -fomit-frame-pointer -D_COMPILE4NIX -c fortran/shadow3.f90 "
-print *,"gfortran   -o ../shadow3 shadow3.o shadow_version.o shadow_globaldefinitions.o stringio.o gfile.o shadow_beamio.o shadow_math.o shadow_variables.o shadow_roughness.o shadow_kernel.o shadow_synchrotron.o shadow_pre_sync.o shadow_pre_sync_urgent.o shadow_preprocessors.o shadow_postprocessors.o shadow_bind_f.o shadow_crl.o "
-print *,"echo 'Note that binary is created in the root directory (./..) '"
-print *,"make[1]: Leaving directory `/users/srio/Oasys/VIRTUAL_ENVIRONMENT/shadow3/src'"
-print *,"+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++"
